Move `command_with_major_version()` to `os` module
