+++ src/os.rs
@@ -1,4 +1,3 @@
-use failure::format_err;
 use semver::Version;
 
 use crate::{settings::PythonVersion, Result};
@@ -21,46 +20,20 @@ pub fn command_with_major_version(
     command: &str,
     interpreter_to_use: &PythonVersion,
 ) -> Result<String> {
-    // NOTE: Make sure the command given by the user contains the major Python version
-    //       appended. This should prevent having a Python 3 interpreter in `.python-version`
-    //       but being called `python` by the user, ending up executing, say, /usr/local/bin/python`
-    //       which is itself a Python 2 interpreter.
-    #[allow(unused_variables)]
-    let last_command_char = format!(
-        "{}",
-        command
-            .chars()
-            .last()
-            .ok_or_else(|| format_err!("Cannot get last character from command {:?}", command))?
-    );
-
-    let command_string_with_major_version = {
-        #[cfg(target_os = "windows")]
-        {
-            // Not implemented yet due to Windows using a `.exe` extension
-            log::error!("Adding the major Python version to binary not implemented on Windows");
-            command.to_string()
-        }
-        #[cfg(not(target_os = "windows"))]
-        {
-            if last_command_char == "2" || last_command_char == "3" {
-                command.to_string()
-            } else {
-                log::debug!(
-                    "Appending Python interpreter major version {} to command.",
-                    interpreter_to_use.version.major
-                );
-                format!("{}{}", command, interpreter_to_use.version.major)
-            }
-        }
-    };
-
-    Ok(command_string_with_major_version)
+    #[cfg(not(target_os = "windows"))]
+    {
+        unix::command_with_major_version(command, interpreter_to_use)
+    }
+    #[cfg(target_os = "windows")]
+    {
+        windows::command_with_major_version(command, interpreter_to_use)
+    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::path::Path;
 
     #[test]
     fn build_filename_from_version_372() {
@@ -74,5 +47,15 @@ mod tests {
         let version = Version::parse("3.7.2-rc1").unwrap();
         let filename = build_filename(&version).unwrap();
         assert!(filename == "Python-3.7.2rc1.tgz" || filename == "python-3.7.2rc1-amd64.exe");
+    }
+
+    #[test]
+    fn append_version_to_command_success() {
+        let interpreter = PythonVersion {
+            location: Path::new("/usr/bin").into(),
+            version: Version::parse("3.7.3").unwrap(),
+        };
+        let cmd = command_with_major_version("python", &interpreter).unwrap();
+        assert_eq!(cmd, "python3");
     }
 }
